+++ snips-nlu-lib/src/errors.rs
@@ -4,14 +4,10 @@ pub enum SnipsNluError {
     ModelLoad(String),
     #[fail(display = "Expected model version {} but found {}", _1, _0)]
     WrongModelVersion(String, &'static str),
-}
-
-#[derive(Debug, Fail)]
-pub enum BuiltinEntityParserError {
-    #[fail(display = "Unable to load builtin entity parser")]
-    LoadingError,
-    #[fail(display = "No builtin entity parser loaded for language '{}'", _0)]
-    ParserNotLoaded(String),
+    #[fail(display = "Unknown intent: '{}'", _0)]
+    UnknownIntent(String),
+    #[fail(display = "Internal error: {}", _0)]
+    InternalError(String)
 }
 
 pub type Result<T> = ::std::result::Result<T, ::failure::Error>;