+++ snips-nlu-lib/src/intent_parser/mod.rs
@@ -1,7 +1,6 @@
 pub mod deterministic_intent_parser;
 pub mod probabilistic_intent_parser;
 
-use std::collections::HashSet;
 use std::path::Path;
 use std::sync::Arc;
 
@@ -13,6 +12,7 @@ use crate::errors::*;
 use crate::models::ProcessingUnitMetadata;
 use crate::resources::SharedResources;
 pub use crate::slot_utils::InternalSlot;
+use crate::utils::IntentName;
 
 pub struct InternalParsingResult {
     pub intent: IntentClassifierResult,
@@ -20,7 +20,7 @@ pub struct InternalParsingResult {
 }
 
 pub fn internal_parsing_result(
-    intent_name: String,
+    intent_name: Option<IntentName>,
     intent_proba: f32,
     slots: Vec<InternalSlot>,
 ) -> InternalParsingResult {
@@ -34,11 +34,11 @@ pub fn internal_parsing_result(
 }
 
 pub trait IntentParser: Send + Sync {
-    fn parse(
-        &self,
-        input: &str,
-        intents: Option<&HashSet<String>>,
-    ) -> Result<Option<InternalParsingResult>>;
+    fn parse(&self, input: &str, intents: Option<&[&str]>) -> Result<InternalParsingResult>;
+
+    fn get_intents(&self, input: &str) -> Result<Vec<IntentClassifierResult>>;
+
+    fn get_slots(&self, input: &str, intent: &str) -> Result<Vec<InternalSlot>>;
 }
 
 pub fn build_intent_parser<P: AsRef<Path>>(