+++ CHANGELOG.md
@@ -2,13 +2,16 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-
 ### Added
+- `get_intents` API: get the probabilities of all intents (including the null intent) with respect to an input text
+- Pass `--top_intents` to the parsing CLI to use the `get_intents` API instead of `parse`
+- `get_slots` API: extract slots by providing a text along with its corresponding intent
 - Added a an optional `CooccurrenceVectorizer` to the `Featurizer` that extracts cooccurrence features
 
 ### Changed
 - Update to Rust 2018
 - Refactored the `Featurizer` and moved its attributes to an underlying `TfidfVectorizer`
+- `Slot` object now contains an optional `confidence_score` attribute
 
 ## [0.61.2] - 2019-01-17
 ### Changed